Clean cpu.c

+++ win/install.sh
@@ -5,9 +5,12 @@ WXMSW=wxMSW-2.8.12
 SDL=SDL-1.2.14
 ZLIB=1.2.3
 
-wget http://diyps3controller.googlecode.com/svn/trunk/win/cpu.c
-gcc -o cpu cpu.c
-rm cpu.c
+if ! test -f cpu
+then
+  wget http://diyps3controller.googlecode.com/svn/trunk/win/cpu.c
+  gcc -o cpu cpu.c
+  rm cpu.c
+fi
 
 CPU=$(./cpu)
